Separate out Scala ForkJoinPool instrumentation from java_concurrent
Finishing task from previous PR.
