+++ dd-java-agent/instrumentation/akka-http-10.0/akka-http-10.0.gradle
@@ -76,9 +76,11 @@ dependencies {
   testCompile group: 'com.squareup.okhttp3', name: 'okhttp', version: '3.6.0'
   testCompile project(':dd-java-agent:instrumentation:trace-annotation')
   testCompile project(':dd-java-agent:instrumentation:akka-concurrent')
+  testCompile project(':dd-java-agent:instrumentation:scala-concurrent')
 
   lagomTestCompile project(':dd-java-agent:instrumentation:akka-http-10.0')
   lagomTestCompile project(':dd-java-agent:instrumentation:akka-concurrent')
+  lagomTestCompile project(':dd-java-agent:instrumentation:scala-concurrent')
   lagomTestCompile group: 'com.lightbend.lagom', name: 'lagom-javadsl-testkit_2.11', version: '1.4.0'
 
   // There are some internal API changes in 10.1 that we would like to test separately for
