+++ dd-java-agent/instrumentation/java-concurrent/java-completablefuture/java-completablefuture.gradle
@@ -1,6 +1,12 @@
+// This is a separate project to allow for testing against various JDKs with existing tooling.
+
 // Set properties before any plugins get loaded
 ext {
   minJavaVersionForTests = JavaVersion.VERSION_1_8
 }
 
 apply from: "${rootDir}/gradle/java.gradle"
+
+dependencies {
+  testCompile project(':dd-java-agent:instrumentation:trace-annotation')
+}
