+++ dd-java-agent/instrumentation/java-concurrent/java-concurrent.gradle
@@ -1,36 +1,11 @@
-// This won't work until the akka and scala integrations are split into separate projects.
-//muzzle {
-//  pass {
-//    coreJdk()
-//  }
-//}
+muzzle {
+  pass {
+    coreJdk()
+  }
+}
 
 apply from: "$rootDir/gradle/java.gradle"
-apply from: "$rootDir/gradle/test-with-scala.gradle"
-
-apply plugin: 'org.unbroken-dome.test-sets'
-
-testSets {
-  slickTest
-}
-
-compileSlickTestGroovy {
-  classpath += files(sourceSets.slickTest.scala.classesDirectory)
-}
 
 dependencies {
-  // This is needed for Scala ForJoinTask/Pool instrumentation
-  compileOnly deps.scala
-
   testCompile project(':dd-java-agent:instrumentation:trace-annotation')
-
-  slickTestCompile project(':dd-java-agent:instrumentation:java-concurrent')
-  slickTestCompile project(':dd-java-agent:instrumentation:trace-annotation')
-  slickTestCompile project(':dd-java-agent:instrumentation:jdbc')
-  slickTestCompile deps.scala
-  slickTestCompile group: 'com.typesafe.slick', name: 'slick_2.11', version: '3.2.0'
-  slickTestCompile group: 'com.h2database', name: 'h2', version: '1.4.197'
 }
-
-// Run Slick library tests along with the rest of unit tests
-test.dependsOn slickTest
