--- dd-java-agent/instrumentation/java-concurrent/scala-testing/scala-testing.gradle
@@ -1,12 +1,0 @@
-apply from: "$rootDir/gradle/java.gradle"
-apply from: "$rootDir/gradle/test-with-scala.gradle"
-
-dependencies {
-  compileOnly deps.scala
-
-  testCompile project(':dd-trace-api')
-  testCompile deps.scala
-
-  testCompile project(':dd-java-agent:instrumentation:java-concurrent')
-  testCompile project(':dd-java-agent:instrumentation:trace-annotation')
-}
