+++ dd-java-agent/instrumentation/play-2.4/play-2.4.gradle
@@ -40,6 +40,7 @@ dependencies {
   testCompile project(':dd-java-agent:instrumentation:netty-4.1')
   testCompile project(':dd-java-agent:instrumentation:akka-http-10.0')
   testCompile project(':dd-java-agent:instrumentation:akka-concurrent')
+  testCompile project(':dd-java-agent:instrumentation:scala-concurrent')
 
   // Before 2.5, play used netty 3.x which isn't supported, so for better test consistency, we test with just 2.5
   testCompile group: 'com.typesafe.play', name: 'play-java_2.11', version: '2.5.0'
