+++ settings.gradle
@@ -105,7 +105,6 @@ include ':dd-java-agent:instrumentation:jax-rs-client-2.0:connection-error-handl
 include ':dd-java-agent:instrumentation:java-concurrent'
 include ':dd-java-agent:instrumentation:java-concurrent:java-completablefuture'
 include ':dd-java-agent:instrumentation:java-concurrent:kotlin-testing'
-include ':dd-java-agent:instrumentation:java-concurrent:scala-testing'
 include ':dd-java-agent:instrumentation:jdbc'
 include ':dd-java-agent:instrumentation:jedis-1.4'
 include ':dd-java-agent:instrumentation:jedis-3.0'
@@ -143,6 +142,7 @@ include ':dd-java-agent:instrumentation:rediscala-1.8.0'
 include ':dd-java-agent:instrumentation:reactor-core-3.1'
 include ':dd-java-agent:instrumentation:rmi'
 include ':dd-java-agent:instrumentation:rxjava-1'
+include ':dd-java-agent:instrumentation:scala-concurrent'
 include ':dd-java-agent:instrumentation:servlet'
 include ':dd-java-agent:instrumentation:servlet:request-2'
 include ':dd-java-agent:instrumentation:servlet:request-3'
